refactor SDK and build
