--- .prettierrc
@@ -1,6 +1,0 @@
-{
-  "semi": false,
-  "singleQuote": true,
-  "arrowParens": "avoid",
-  "trailingComma": "none"
-}
