+++ assets/front_logo@2x.png
