+++ examples/expo/.gitignore
@@ -1,5 +1,5 @@
 node_modules/
-.expo/
+../.expo/
 dist/
 npm-debug.*
 *.jks
