+++ examples/expo/assets/adaptive-icon.png
