+++ examples/expo/assets/favicon.png
