+++ examples/expo/assets/icon.png
