+++ examples/expo/assets/logo.png
