+++ examples/expo/assets/splash.png
