--- packages/front-b2b-link-rn/package.json
@@ -1,37 +1,0 @@
-{
-  "name": "@front-finance/frontfinance-rn-sdk",
-  "version": "1.2.1",
-  "description": "Front Finance React Native SDK.",
-  "main": "./src/index.ts",
-  "module": "./src/index.ts",
-  "types": "./src/index.ts",
-  "type": "module",
-  "license": "MIT",
-  "author": "Front Finance, Inc",
-  "repository": "https://github.com/FrontFin/front-b2b-link-rn.git",
-  "homepage": "https://github.com/FrontFin/front-b2b-link-rn",
-  "devDependencies": {
-    "@types/react": "^18.2.15",
-    "react": "^18.2.0",
-    "react-native": "^0.72.3"
-  },
-  "peerDependencies": {
-    "react": "^18.2.0",
-    "react-native": "^0.72.3"
-  },
-  "dependencies": {
-    "react-native-webview": "11.26.0"
-  },
-  "scripts": {
-    "build": "rimraf dist && yarn build:esm && yarn build:cjs && yarn copy",
-    "build:esm": "tsc --outDir dist",
-    "build:cjs": "tsc --module commonjs --outDir dist/cjs",
-    "typecheck": "tsc --noEmit",
-    "lint": "eslint --ext .js,.ts,.tsx .",
-    "copy": "node ./tools/copy",
-    "publish:npm": "yarn build && cd dist && yarn publish --access public"
-  },
-  "keywords": [
-    "front-finance"
-  ]
-}
