--- packages/front-b2b-link-rn/tsconfig.json
@@ -1,22 +1,0 @@
-{
-  "compilerOptions": {
-    "module": "commonjs",
-    "target": "es6",
-    "outDir": "./dist",
-    "strict": true,
-    "moduleResolution": "node",
-    "esModuleInterop": true,
-    "skipLibCheck": true,
-    "forceConsistentCasingInFileNames": true,
-    "noEmit": false,
-    "declaration": true,
-    "jsx": "react-native",
-    "lib": ["es6"],
-    "baseUrl": "./",
-    "paths": {
-      "*": ["node_modules/*", "src/types/*"]
-    }
-  },
-  "include": ["src"],
-  "exclude": ["node_modules", "__tests__"]
-}
