+++ src/index.ts
@@ -1,3 +1,2 @@
 import FrontFinance from './Front'
-export * from './Types'
 export { FrontFinance }
